add bluetoothAddress to SimpleBLE namespace and remove external dependency
